requirements.txt
langgraph
langchain
langsmith
langchain-google-genai
langchain-groq

fastapi
uvicorn

sentence-transformers
--extra-index-url https://download.pytorch.org/whl/cpu
torch

psycopg2-binary
sqlalchemy
neo4j
chromadb

pandas
tqdm
python-dotenv
pydantic

trustcall
#db sync
psycopg2
confluent_kafka